# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAP API",
    "version": "1.0.0",
    "description": "Kontrak rancangan SAP untuk frontend/backend. Kebutuhan otorisasi dan domain ada di API_SPEC.md. Belum merupakan layanan berjalan."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/users/me": {
      "patch": {
        "operationId": "updateMe",
        "summary": "Ubah display name",
        "tags": [
          "users"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "409": {
            "description": "Revision or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "413": {
            "description": "Image too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "415": {
            "description": "Unsupported image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Domain validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMe",
        "summary": "Antrekan penghapusan dan berikan receipt cookie",
        "tags": [
          "users"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Satu key per aksi; retry payload sama memakai key sama selama24jam."
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted asynchronous work",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "409": {
            "description": "Revision or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "413": {
            "description": "Image too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "415": {
            "description": "Unsupported image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Domain validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteInput"
              }
            }
          }
        }
      }
    },
    "/users/me/preferences": {
      "patch": {
        "operationId": "updatePreferences",
        "summary": "Simpan preferensi akun (sakelar SAPA)",
        "tags": [
          "users"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferencesInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Domain validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "user"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Meta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "sapaEnabled": {
            "type": "boolean",
            "description": "Preferensi teman virtual SAPA. Default true untuk akun baru; UI membaca ini saat bootstrap. Lihat addendum SAPA_ASSISTANT.md."
          }
        },
        "required": [
          "id",
          "displayName",
          "email",
          "role",
          "emailVerified",
          "sapaEnabled"
        ],
        "additionalProperties": false
      },
...
      "ProfileInput": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          }
        },
        "required": [
          "displayName"
        ],
        "additionalProperties": false
      },
...
      "UserPreferences": {
        "type": "object",
        "properties": {
          "sapaEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "sapaEnabled"
        ],
        "additionalProperties": false
      },
      "UserPreferencesInput": {
        "type": "object",
        "description": "Hanya field preferensi yang boleh diubah. userId/role diambil dari session, tidak dari body.",
        "properties": {
          "sapaEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "sapaEnabled"
        ],
        "additionalProperties": false
      },
...
      "DeleteInput": {
        "type": "object",
        "properties": {
          "confirmation": {
            "type": "string",
            "enum": [
              "DELETE_MY_ACCOUNT"
            ]
          }
        },
        "required": [
          "confirmation"
        ],
        "additionalProperties": false
      },
      "Deletion": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ]
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "requestId",
          "status",
          "requestedAt",
          "completedAt"
        ],
        "additionalProperties": false
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/User"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "UserPreferencesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserPreferences"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "DeletionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Deletion"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sap_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "deletionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sap_deletion"
      }
    },
    "headers": {
      "ContractVersion": {
        "description": "Contract version echoed on every response (success and error) so clients can detect contract drift.",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "1.0.0"
          ]
        }
      },
      "SetCookie": {
        "description": "Sets or clears a host-only cookie. All auth cookies are HttpOnly, Secure, SameSite=Lax, Path=/. sap_csrf carries the signed double-submit CSRF token (GET /auth/csrf; rotated on login/logout). sap_session is the opaque session token set on verify-email/login and cleared on logout, reset-password, and account deletion. sap_deletion is the account-deletion receipt set by DELETE /users/me. A single response may emit more than one Set-Cookie line.",
        "schema": {
          "type": "string"
        }
      }
    }
  }
}